Create README.md
This project allows people to sign into their accounts to add new bookmarks and to view the bookmarks they already added. The home page is visible to all and shows the five most popular bookmarks.

The backend code is written in C# using the ASP.NET Core with React framework. The frontend is written in React. The database is SQL Server using Entity Framework to connect it to the backend code.